+++ src/pages/user/HomePage/ProductRecommend.js
@@ -1,0 +1,63 @@
+import { useState } from "react";
+import {
+  Product,
+  ProductLink,
+  ProductImage,
+  Pointer,
+  ProductName,
+  RecommendContent,
+  BlankCard,
+} from "./style";
+
+function getRecommendProduct() {
+  return fetch("/api/category/product", {
+    method: "GET",
+  })
+    .then((res) => res.json())
+    .then((res) => {
+      const recommendProducts = res.data.filter(
+        (catergory) => catergory.name === "人氣商品"
+      );
+      return recommendProducts;
+    })
+    .then((res) => {
+      return res[0].Products;
+    });
+}
+
+function RecommendItem(props) {
+  return (
+    <Product>
+      <ProductLink to={props.itemLink}>
+        <ProductImage src={props.itemImg}></ProductImage>
+        <Pointer>
+          <span>➜</span>
+        </Pointer>
+        <ProductName>{props.itemName}</ProductName>
+      </ProductLink>
+    </Product>
+  );
+}
+
+export default function ProductRecommend() {
+  const [products, setProducts] = useState([]);
+
+  getRecommendProduct().then((res) => {
+    setProducts(res);
+  });
+
+  return (
+    <RecommendContent>
+      {products &&
+        products.map((product) => (
+          <RecommendItem
+            itemLink={`/product/${product.id}`}
+            itemImg={product.image}
+            itemName={product.name}
+          />
+        ))}
+      <BlankCard />
+      <BlankCard />
+    </RecommendContent>
+  );
+}
